pyproject.toml
[tool.poetry]
name = "transvib"
version = "0.1.0"
description = ""
authors = ["gunnar.kistner@gmail.com <gunnar.kistner@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.85.1"
uvicorn = {extras = ["standard"], version = "^0.19.0"}
strawberry-graphql = {extras = ["fastapi"], version = "^0.137.1"}
tortoise-orm = "^0.19.2"
python-jose = "^3.3.0"
passlib = "^1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
